Addtional types for page view updates column
