+++ app/graphql/mutations/create_page_view.rb
@@ -7,7 +7,7 @@ module Mutations
     argument :template, String, required: true
     argument :page_type, String, required: false
     argument :url, String, required: true
-    # argument :mutations,[Types::SiteViewMutationInputType],required: true
+    argument :mutations,[Types::SiteViewMutationInputType],required: false
     argument :site_id, Integer, required: true
 
     def resolve(args)
@@ -17,15 +17,17 @@ module Mutations
       end
       page_view = site.page_views.new()
       page_view.attributes = args.slice(:title,:template,:url,:page_type)
-      # mutations = args[:mutations].clone.map do |mutation|
-      #   begin
-      #     mutation[:payload] = JSON.parse(mutation[:payload])
-      #   rescue StandardError # rubocop:disable Lint/HandleExceptions
-      #     # use payload as string if it's not a json
-      #   end
-      #   mutation.to_h
-      # end
-      # page_view.updates = SiteViewUpdaterService.compact(view.mutations + mutations)
+      if args[:mutations]
+        mutations = args[:mutations].clone.map do |mutation|
+          begin
+            mutation[:payload] = JSON.parse(mutation[:payload])
+          rescue StandardError # rubocop:disable Lint/HandleExceptions
+            # use payload as string if it's not a json
+          end
+          mutation.to_h
+        end
+        page_view.updates = SiteViewUpdaterService.compact(page_view.mutations + mutations)
+      end
       if page_view.save
         { page_view: page_view, errors: nil }
       else
