+++ app/graphql/mutations/update_page_view.rb
@@ -7,21 +7,23 @@ module Mutations
     argument :template, String, required: false
     argument :page_type, String, required: false
     argument :url, String, required: false
-    # argument :mutations,[Types::SiteViewMutationInputType],required: true
+    argument :mutations,[Types::SiteViewMutationInputType],required: false
     argument :id, Integer, required: true
 
     def resolve(args)
       page_view = page_view(args[:id])
       page_view.attributes = args.slice(:title,:template,:url,:page_type)
-      # mutations = args[:mutations].clone.map do |mutation|
-      #   begin
-      #     mutation[:payload] = JSON.parse(mutation[:payload])
-      #   rescue StandardError # rubocop:disable Lint/HandleExceptions
-      #     # use payload as string if it's not a json
-      #   end
-      #   mutation.to_h
-      # end
-      # page_view.updates = SiteViewUpdaterService.compact(view.mutations + mutations)
+      if args[:mutations]
+        mutations = args[:mutations].clone.map do |mutation|
+          begin
+            mutation[:payload] = JSON.parse(mutation[:payload])
+          rescue StandardError # rubocop:disable Lint/HandleExceptions
+            # use payload as string if it's not a json
+          end
+          mutation.to_h
+        end
+        page_view.updates = SiteViewUpdaterService.compact(page_view.mutations + mutations)
+      end
       if page_view.save
         { page_view: page_view, errors: nil }
       else
