+++ app/graphql/types/page_view_config_type.rb
@@ -1,0 +1,11 @@
+module Types
+  class PageViewConfigType < Types::BaseObject
+    #Empty string as temporarily to be able to call type
+    field :blank,String, null:false
+
+
+    def blank
+      ""
+    end
+  end
+end
