+++ app/graphql/types/page_view_type.rb
@@ -5,5 +5,7 @@ module Types
     field :template, String, null:false
     field :url, String, null:false
     field :page_type, String, null:false
+    # Change the name of this most likely
+    field :updates, PageViewConfigType, null:true
   end
 end
