+++ app/models/page_view.rb
@@ -2,4 +2,8 @@ class PageView < ApplicationRecord
   belongs_to :site
   validates :site, presence: true
   enum page_type: [ :study ]
+  include MutationHelpers
+
+
+
 end
