+++ app/models/site_view.rb
@@ -9,6 +9,7 @@ DEFAULT_AGG_ORDER = {
 
 class SiteView < ApplicationRecord # rubocop:disable Metrics/ClassLength
   belongs_to :site
+  include MutationHelpers
   before_save do
     if default_changed? && default
 
@@ -37,17 +38,7 @@ class SiteView < ApplicationRecord # rubocop:disable Metrics/ClassLength
     end
   end
 
-  def mutations
-    updates.map do |update|
-      mutation = update.clone.deep_symbolize_keys
-      begin
-        mutation[:payload] = JSON.parse(mutation[:payload], quirks_mode: true)
-      rescue StandardError # rubocop:disable Lint/HandleExceptions
-        # use payload as string if it's not a json
-      end
-      mutation
-    end
-  end
+  
 
   def all_fields # rubocop:disable Metrics/MethodLength
     %w[
